requirements.txt
# Entferne textract, da es manuell im Dockerfile installiert wird
# textract==1.6.5  # Entfernt

# Liste aller anderen benötigten Pakete
accelerate==1.1.1
aioboto3==13.2.0
aiobotocore==2.15.2
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioitertools==0.12.0
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
anytree
argcomplete
asttokens
attrs==24.2.0
autograd==1.7.0
beartype==0.18.5
beautifulsoup4
blinker==1.9.0
boto3==1.35.36
botocore==1.35.36
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
chardet  # Direkt von textract entfernt, falls nicht benötigt
charset-normalizer==3.4.0
click==8.1.7
compressed_rtf==1.0.6
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
deepdiff==8.0.1
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
docx2txt
executing==2.1.0
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
graspologic==3.4.1
graspologic-native==1.2.1
h11==0.14.0
hnswlib==0.8.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
hyppo==0.4.0
identify==2.6.2
idna==3.10
IMAPClient
iniconfig==2.0.0
instructor
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpickle==4.0.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langchain
langchain-community
langchain-core==0.2.29
langchain-openai
langchain-text-splitters
langchain-unstructured==0.1.1
langsmith==0.1.142
llvmlite==0.43.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nano-vectordb==0.0.4.3
narwhals==1.13.4
neo4j==5.26.0
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
numba==0.60.0
numpy==1.26.4
olefile==0.47
ollama==0.3.3
openai==1.54.3
oracledb==2.5.0
orderly-set==5.2.2
orjson==3.10.11
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
POT==0.9.5
pre_commit==4.0.1
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.2.0
pypdf==5.1.0
pypdfium2==4.30.0
pytest==8.3.3
python-dateutil
python-dotenv==1.0.1
python-pptx
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
s3transfer==0.10.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.12.0
seaborn==0.13.2
setuptools==75.4.0
shellingham==1.5.4
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.8.1
SQLAlchemy==2.0.35
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.40.1
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
toml==0.10.2
torch==2.5.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
typer==0.13.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
umap-learn==0.5.7
urllib3==2.2.3
virtualenv==20.27.1
wcwidth==0.2.13
wheel
wrapt==1.16.0
xlrd
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.17.1